feat: Reorganize metrics commands with subcommand structure
Implements Issue #4 by reorganizing metrics-related commands under a single 'metrics' command with subcommands:

• Base command: `metrics` - Shows code quality overview (existing behavior)
• Subcommands:
  - `metrics collect` - Collect metrics for tracking (was collect-metrics)
  - `metrics show` - Show function history (was show-metrics)
  - `metrics trends` - Analyze trends (was analyze-trends)
  - `metrics pr` - Show PR metrics (was pr-metrics)

Changes:
- Create src/commands/metrics.ts with organized subcommand structure
- Update src/cli.ts to use new metrics command and remove old individual commands
- Maintain full backward compatibility for base metrics functionality
- Update CLAUDE.md documentation with new command structure
- Add comprehensive test coverage for reorganized commands

Benefits:
- Cleaner command organization under logical grouping
- Improved discoverability with --help structure
- Maintains existing functionality while improving UX
- Prepares for Phase 2 feature organization goals
